make 8-bit ogg theora

+++ histutils/normframe.py
@@ -1,4 +1,5 @@
 from __future__ import division
+from numpy import float32 #memory saving
 
 """
 inputs:
@@ -10,4 +11,4 @@ Clim: length 2 of tuple or numpy 1-D array specifying lowest and highest expecte
 def normframe(I,Clim):
     Vmin = Clim[0]; Vmax = Clim[1]
 
-    return (I.astype('float64').clip(Vmin, Vmax) - Vmin) / (Vmax - Vmin) #stretch to [0,1]
+    return (I.astype(float32).clip(Vmin, Vmax) - Vmin) / (Vmax - Vmin) #stretch to [0,1]
